// apps/server/src/models/youtube.types.ts
// YouTube API types

export interface VideoInfo {
  videoId: string;
  title: string;
  publishedAt: string;
}

export interface VideoDetails {
  title: string;
  channelId: string;
  publishedAt: string;
  thumbnails: any;
  tags?: string[];
  description?: string;
  durationSeconds?: number;
  privacyStatus?: string;
}

export interface ChannelInfo {
  channelId: string;
  title?: string;
}